Apply dhewm3 / 0c84e4bff3ee8db9a6b8cb89af39ef2870ec0b90: Fix shared library loading for x86_64 Library handles are "void *" on posix.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6192 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ framework/Common.cpp
@@ -216,7 +216,7 @@ private:
 	idStrList					warningList;
 	idStrList					errorList;
 
-	int							gameDLL;
+    uintptr_t					gameDLL;
 
 #ifdef ID_WRITE_VERSION
 	idCompressor *				config_compressor;
